requirements.txt
absl-py==0.11.0
asgiref==3.3.1
astor==0.8.1
astunparse==1.6.3
beautifulsoup4==4.9.3
bokeh==2.2.3
Brotli==1.0.9
cached-property==1.5.2
cachetools==4.1.1
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cloudpickle==1.6.0
colorcet==2.0.6
contextvars==2.4
cryptography==3.2.1
cycler==0.10.0
dash==1.20.0
dash-core-components==1.16.0
dash-html-components==1.1.3
dash-renderer==1.9.1
dash-table==4.11.3
dask==2021.2.0
dataclasses==0.8
datashader==0.12.0
datashape==0.5.2
decorator==4.4.2
defusedxml==0.6.0
Deprecated==1.2.10
distributed==2021.2.0
Django==3.0.5
django-allauth==0.44.0
django-appconf==1.0.4
django-plotly-dash==1.6.3
djongo==1.3.4
dpd-components==0.1.0
Flask==1.1.2
Flask-Compress==1.9.0
flatbuffers==1.12
fsspec==0.8.5
future==0.18.2
gast==0.3.3
google-api-core==1.23.0
google-api-python-client==1.12.8
google-auth==1.23.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
googleapis-common-protos==1.52.0
grpcio==1.32.0
h5py==2.10.0
HeapDict==1.0.1
holoviews==1.14.1
httplib2==0.18.1
idna==2.10
imageio==2.9.0
immutables==0.15
importlib-metadata==3.4.0
itsdangerous==1.1.0
Jinja2==2.11.3
jira==2.0.0
joblib==1.0.0
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
llvmlite==0.35.0
locket==0.2.1
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.4
msgpack==1.0.2
multipledispatch==0.6.0
networkx==2.5
nltk==3.5
numba==0.52.0
numpy==1.19.5
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.9
pandas==1.1.5
panel==0.10.3
param==1.10.1
partd==1.1.0
pbr==5.5.1
Pillow==8.1.0
pkg-resources==0.0.0
plotly==4.14.3
protobuf==3.14.0
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyct==0.4.8
PyGithub==1.54
PyJWT==1.7.1
pymongo==3.11.2
pynndescent==0.5.2
pyparsing==2.4.7
PyQt5==5.15.2
PyQt5-sip==12.8.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.4
pyviz-comms==2.0.1
PyWavelets==1.1.1
PyYAML==5.4.1
regex==2020.11.13
requests==2.24.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
retrying==1.3.3
rsa==4.6
scikit-image==0.17.2
scikit-learn==0.24.1
scipy==1.5.4
six==1.15.0
sklearn==0.0
sortedcontainers==2.3.0
soupsieve==2.1
sqlparse==0.2.4
tblib==1.7.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-addons==0.12.0
tensorflow-estimator==2.4.0
termcolor==1.1.0
threadpoolctl==2.1.0
tifffile==2020.9.3
toolz==0.11.1
tornado==6.1
tqdm==4.56.2
typeguard==2.10.0
typing-extensions==3.7.4.3
umap==0.1.1
umap-learn==0.5.1
uritemplate==3.0.1
urllib3==1.25.11
Werkzeug==1.0.1
wrapt==1.12.1
xarray==0.16.2
zict==2.0.0
zipp==3.4.0
